+++ lua/js-i18n/translation_source.lua
